MINIFICPP-264: CompressContent Processor
This closes #151.

+++ libminifi/include/io/AtomicEntryStream.h
@@ -61,7 +61,7 @@ class AtomicEntryStream : public BaseStream {
    */
   void seek(uint64_t offset);
 
-  virtual const uint32_t getSize() const {
+  virtual const uint64_t getSize() const {
     return length_;
   }
